+++ CHANGES.md
@@ -3,3 +3,4 @@
 ## v0.1.1
 
 - fix(deps): update dependency `js-yaml` to version `4.1.1`.
+- fix(algo): throw detailed error on algorithm mismatch in EdDSA verifier.